@@ -1,13 +1,10 @@
-var fs = require('fs');
-var ServiceProvider = require('saml2-js').ServiceProvider;
+var { ServiceProvider } = require('saml2-js');
 
 const serviceProvider = new ServiceProvider({
-    entity_id: 'saml-prod',
-    private_key: fs.readFileSync(process.cwd() + '\\src\\certs\\key.pem').toString(),
-    certificate: fs.readFileSync(process.cwd() + '\\src\\certs\\cert.pem').toString(),
-    assert_endpoint: process.env.baseUrl + 'api/auth/login/saml',
+    entity_id: process.env.KEYCLOAK_CLIENT_ID,
+    assert_endpoint: process.env.BASE_URL + 'api/auth/signin',
     allow_unencrypted_assertion: true,
     notbefore_skew: 19800,
 });
 
-module.exports = serviceProvider;
+module.exports = { serviceProvider };
