@@ -60,9 +60,16 @@ export default function ScannerListContaner({...props}) {
         )
         .then(res => {
             setPageIndex(1);
-            mutate(`${process.env.backendUrl}api/scanners?page=${pageIndex}&limit=${rowsPerPage}&sort=-lastActive`);
-            setLoadingRemove(false);
-            setRemoveScanner(false);
+            mutate(`${process.env.backendUrl}api/scanners?page=${pageIndex}&limit=${rowsPerPage}&sort=-lastActive`)
+                .then(() => {
+                    setLoadingRemove(false);
+                    toast.success('Successfully deleted scanner');
+                    setRemoveScanner(false);
+                }, () => {
+                    setLoadingRemove(false);
+                    toast.error('Failed to delete scanner');
+                    setRemoveScanner(false);
+                });
         })
         .catch(err => {
             toast.error('Failed to delete scanner');
@@ -77,12 +84,16 @@ export default function ScannerListContaner({...props}) {
 
     return (
         <>
-            <Card withpadding onClick={handleScannerDetailClick}>
+            <Card 
+                withpadding 
+                onClick={handleScannerDetailClick} 
+                hover
+            >
                 <Grid container spacing={2}>
                 <Grid item xs={12} display='flex' justifyContent="space-between">
                     <Stack direction='row' width={0.8} spacing={1}>
                     <Box display='flex' alignItems='center'>
-                        <Image src="/Vectorscanner.png" layout="fixed" width={43} height={32} />
+                        <Image alt={"Scanner"} src="/Vectorscanner.png" layout="fixed" width={43} height={32} />
                     </Box>
                     <Box maxWidth={1}>
                         <Typography noWrap sx={{ fontWeight: 500, fontSize: '20px', lineHeight: '24px', color: '#190D29' }}>
