@@ -20,19 +20,17 @@ export default function Step2() {
             method: "POST",
             data: { registrationToken: registrationToken },
         })
-            .then(async (data) => {
+            .then((data) => {
                 setScannerData(data);
+                destroyCookie({}, authConstants.REGISTRATION_TOKEN, {path: '/'});
+                destroyCookie({}, authConstants.SESSION_TOKEN, { path: '/scanners' });
+                destroyCookie({}, authConstants.SESSION_TOKEN, { path: '/' });
                 setStep(3);
                 // console.log(data);
             })
             .catch((err) => {
                 setError(err.response?.data?.message || err.message);
             })
-
-        destroyCookie({}, authConstants.REGISTRATION_TOKEN);
-        destroyCookie({}, authConstants.CALLBACK_URL);
-        destroyCookie({}, authConstants.SESSION_TOKEN);
-        destroyCookie({}, authConstants.CSRF_TOKEN);
     }
     
     useEffect(() => {
