@@ -2,6 +2,6 @@ import Image from "next/image";
 
 export default function SettingsIcon() {
     return (
-        <Image src="/icons/Vectorsettings.png" width={20} height={20} />
+        <Image alt={"Settings"} src="/icons/Vectorsettings.png" width={20} height={20} />
     )
 }
