@@ -1,39 +1,55 @@
 import Register from "components/PageComponents/RegisterScanner";
-import { parseCookies, setCookie } from "nookies";
+import { destroyCookie, parseCookies, setCookie } from "nookies";
 import { authConstants } from "constants/auth";
-import jwt from 'jsonwebtoken';
 import { RegisterProvider } from "lib/contexts/registerContext";
 
 export default ({...props}) => <RegisterProvider {...props} >
         <Register />
     </RegisterProvider>;
 
-export function getServerSideProps(ctx) {
-    const { registrationToken } = ctx.query;
-    setCookie(ctx, authConstants.REGISTRATION_TOKEN, registrationToken, {
-        path: '/'
-    });
+export async function getServerSideProps(ctx) {
+    const { registrationToken, callback } = ctx.query;
 
-    // Check if user is authorized
-    const token = parseCookies(ctx)[authConstants.SESSION_TOKEN];
+    var token = parseCookies(ctx)[authConstants.SESSION_TOKEN];
+
+    if (typeof callback === 'undefined') {
+        destroyCookie(ctx, authConstants.SESSION_TOKEN,{
+            path: '/'
+        });
+        token = null;
+    }
+    if (registrationToken && registrationToken !== 'undefined') {
+        setCookie(ctx, authConstants.REGISTRATION_TOKEN, registrationToken, {
+            path: '/',
+            sameSite: 'lax'
+        });
+    } else {
+        ctx.res.writeHead(302, {
+            Location: "/dashboard",
+        });
+  
+      return ctx.res.end();
+    }
+    
     if (token) {
-        try {
-            const verifiedUser = jwt.verify(token, process.env.JWT_SECRET);
-            return {
-                props: {
-                    user: verifiedUser
-                }
-            };
-        } catch (e) {
-            // User not authorized
-            return {
-                props: {
-                    user: null
-                }
-            };
+        // Check if user is authorized
+        const { verified } = await fetch(`${process.env.backendUrl}api/auth/verify`, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({token: token})
+        }).then(res => res.json());
+        return {
+            props: {
+                user: verified
+            }
         }
     }
+
     return {
-        props: {}
+        props: {
+            user: false
+        }
     };
 }
