@@ -1,9 +1,10 @@
 export const getRouteTitle = (path) => {
   switch(path) {
+    case '/setting':
+      return 'Scanner Settings';
     case '/dashboard':
       return 'Scanbot Dashboard';
     case '/api/auth/login/saml':
-    case '/api/auth/bypass/saml':
       return 'Scanbot Login';
     case '/api/auth/logout/saml':
       return 'Scanbot Logout';
